scraper-service/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
oracledb==2.1.1
python-dotenv==1.0.0
structlog==23.2.0
pytest==7.4.3